python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
mysql-connector-python==8.4.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
